+++ R/Agenerics.R
@@ -160,6 +160,7 @@ if (!isGeneric("ncol")) { setGeneric("nrow", function(x) standardGeneric("nrow")
 if (!isGeneric("ncol<-")) { setGeneric("ncol<-", function(x, ..., value) standardGeneric("ncol<-")) }
 if (!isGeneric("nrow<-")) { setGeneric("nrow<-", function(x, ..., value) standardGeneric("nrow<-")) }
 if (!isGeneric("origin")) {	setGeneric("origin", function(x, ...) standardGeneric("origin")) }
+if (!isGeneric("origin<-")) {setGeneric("origin<-", function(x, value)	standardGeneric("origin<-"))}
 if (!isGeneric("pairs")) { setGeneric("pairs", function(x, ...)	standardGeneric("pairs"))}
 if (!isGeneric("patches")) {setGeneric("patches", function(x, ...) standardGeneric("patches"))}
 if (!isGeneric("persp")) { setGeneric("persp", function(x,...) standardGeneric("persp")) }