+++ src/raster_stats.cpp
@@ -440,7 +440,7 @@ SpatDataFrame SpatRaster::zonal(SpatRaster z, std::string fun, bool narm, SpatOp
 		out.setError("zonal SpatRaster has no values");
 		return(out);
 	}
-	if (!compare_geom(z, false, true)) {
+	if (!compare_geom(z, false, true, opt.get_tolerance())) {
 		out.setError("dimensions and/or extent do not match");
 		return(out);
 	}